refactor: ParameterGroup 클래스 제거 및 cascade 옵션 변경

+++ backend/project/src/main/java/com/sk/skala/axcalibur/feature/dto/ApiInfoDTO.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ApiInfoDto {
-    private Long apiPk;         // API PK
+    private Integer apiPk;         // API PK
     private String id;          // API 아이디
     private String name;        // API 이름
     private String desc;        // API 설명
